refactor: extract emulation cloud restore logic into a standalone reusable modal component
